[DMS-1124] Slice 5 Findings: trim test-only production surface in profile merge contracts
Reduce test-only production API in the profile merge implementation while
preserving full behavior coverage. Two clear wins from the audit; a third
(walker test-only index accessors) was deferred as out of scope.

CurrentCollectionRowProjection / CurrentSeparateScopeRowProjection:
- Drop the ParentPhysicalIdentityValues field from both projection records.
  Production code constructs but never reads it; the doc-comment described it
  as redundant with the index key. Three production construction sites
  (ProfileCollectionWalker.BuildCurrentCollectionRowsIndex,
  ProfileCollectionWalker.BuildCurrentSeparateScopeRowsIndex,
  RelationalWriteProfileMerge.BuildCurrentSeparateScopeRowProjection) lose the
  trailing argument.
- RelationalWriteProfileMerge.BuildCurrentSeparateScopeRowProjection no longer
  takes parentPhysicalIdentityValues; the caller drops the now-unused argument.
- Test reads migrated:
  - ProfileCollectionWalkerTests line 125 redundant index-key assertion deleted.
  - Two test sites in the aligned-scope dispatch fixture redirected to the
    AlignedScopeInvocation capture record's already-captured
    ParentPhysicalIdentityValues, asserting on what the synthesizer was actually
    invoked with rather than what the projection record carried.
- Three test construction sites in SynthesizeSeparateScopeInstanceTests drop
  the trailing argument.

CollectionWriteCandidate.InferSemanticIdentityInOrderForTests:
- Move the lossy presence-inference helper out of production
  RelationalWriteContracts.cs into a new test-project helper class
  EdFi.DataManagementService.Backend.Tests.Unit.SemanticIdentityTestHelpers
  with the same signature. Production now has zero references to the inference
  algorithm; the helper lives where it is exclusively consumed.
- Update the CollectionWriteCandidate constructor argument-error message: drop
  the nameof(InferSemanticIdentityInOrderForTests) reference (which would
  otherwise dangle once the production method is removed) and replace it with a
  prose mention of the JSON-side presence probes the production rule requires.
- Update ~40 call sites across 7 test files (RelationalWriteContractsTests,
  RelationalWriteNoProfileMergeSynthesizerTests, SynthesizeSeparateScopeInstance,
  RelationalWriteProfileMergeSynthesizer, ProfileCollectionWalker,
  ProfileCollectionWalkerUnconsumedCandidate, ProfileCollectionMatchedRowOverlay)
  from CollectionWriteCandidate.InferSemanticIdentityInOrderForTests(...) to
  SemanticIdentityTestHelpers.InferSemanticIdentityInOrderForTests(...).
- Update one test exception-message wildcard pattern that previously asserted
  on the nameof token to assert on the new prose token instead.

Walker test-only index accessors deferred (audit only):
- The 5 internal For-testing-only accessors on ProfileCollectionWalker
  (CurrentCollectionRowsByTableAndParentIdentity,
  CurrentSeparateScopeRowsByTableAndParentIdentity,
  VisibleStoredRowsByChildScopeAndParent,
  VisibleRequestItemsByChildScopeAndParent,
  CurrentRowsByJsonScopeAndParent) are read at ~30 sites across 16 fixtures in
  a 7,500-line test file. Narrowing options considered: (a) extract a
  ProfileWalkerIndexes collaborator the walker composes and tests construct
  separately, (b) bundle the 5 accessors into a single IndexesForTesting
  property, (c) replace per-partition isolation assertions with WalkChildren
  outcomes. (a) and (b) are mechanical churn for cosmetic gain; (c) loses the
  per-partition-not-per-scope invariants those tests pin (e.g. that parent A's
  current-row bucket cannot contain parent B's children). Unlike the
  ParentPhysicalIdentityValues dead field and the inference algorithm, these
  accessors only expose state behind already-internal visibility and are
  documented as test-only. Out of this task scope per Keep-changes-scoped.

Verification:
- dotnet build clean for backend production project and unit test project.
- dotnet csharpier format clean on all 13 touched files.
- dotnet test: 1150 passed, 0 failed.
