+++ src/dms/backend/EdFi.DataManagementService.Backend.Tests.Unit/Profile/ProfileCollectionWalkerUnconsumedCandidateTests.cs
@@ -383,7 +383,7 @@ internal static class UnsupportedTopologyBuilders
             requestOrder: requestOrder,
             values: values,
             semanticIdentityValues: [identityValue],
-            semanticIdentityInOrder: CollectionWriteCandidate.InferSemanticIdentityInOrderForTests(
+            semanticIdentityInOrder: SemanticIdentityTestHelpers.InferSemanticIdentityInOrderForTests(
                 orphanPlan,
                 [identityValue]
             )
